+++ projects/extension/CHANGELOG.md
@@ -2,9 +2,20 @@
 
 ## [Unreleased]
 
+- Add "latest block" information to Popup, ([#1254](https://github.com/paritytech/substrate-connect/pull/1254))
+
+## 0.1.10 - 2022-09-15
+
 ### Changed
 
-- Update @substrate/smoldot-light to [version 0.6.30](https://github.com/paritytech/smoldot/blob/main/bin/wasm-node/CHANGELOG.md#0630---2022-08-12). ([#1213](https://github.com/paritytech/substrate-connect/pull/1213))
+- Update @substrate/smoldot-light to [version 0.6.33](https://github.com/paritytech/smoldot/blob/main/bin/wasm-node/CHANGELOG.md#0633---2022-09-13). ([#1248](https://github.com/paritytech/substrate-connect/pull/1248))
+
+## 0.1.9 - 2022-09-12
+
+### Changed
+
+- Show error message when smoldot crashes in both Popup and Options page. ([#1128](https://github.com/paritytech/substrate-connect/pull/1228))
+- Update @substrate/smoldot-light to [version 0.6.32](https://github.com/paritytech/smoldot/blob/main/bin/wasm-node/CHANGELOG.md#0632---2022-09-07). ([#1244](https://github.com/paritytech/substrate-connect/pull/1244))
 
 ## 0.1.8 - 2022-07-12